Refactor cypress dir
