+++ package.json
@@ -53,7 +53,7 @@
     "@types/react-test-renderer": "^16.9.1",
     "@types/sanitize-filename": "^1.6.3",
     "antlr4ts-cli": "^0.5.0-alpha.4",
-    "cypress": "^13.6.1",
+    "cypress": "^12.8.1",
     "cypress-watch-and-reload": "^1.10.6",
     "eslint": "^6.8.0",
     "husky": "6.0.0",
